requirements.txt
alembic==1.13.1
annotated-types==0.6.0
anyio==4.3.0
asgi-lifespan==2.1.0
async-asgi-testclient==1.4.11
async-timeout==4.0.3
asyncpg==0.29.0
bcrypt==4.1.2
certifi==2024.2.2
cffi==1.16.0
charset-normalizer==3.3.2
click==8.1.7
cryptography==42.0.4
fastapi==0.109.2
greenlet==3.0.3
h11==0.14.0
idna==3.6
iniconfig==2.0.0
Mako==1.3.2
MarkupSafe==2.1.5
multidict==6.0.5
packaging==23.2
pluggy==1.4.0
psycopg2-binary==2.9.9
pycparser==2.21
pydantic==2.6.1
pydantic_core==2.16.2
PyJWT==2.8.0
pytest==8.0.1
pytest-asyncio==0.23.5
python-dotenv==1.0.1
requests==2.31.0
sniffio==1.3.0
SQLAlchemy==2.0.27
SQLAlchemy-Utils==0.41.1
starlette==0.36.3
typing_extensions==4.9.0
urllib3==2.2.1
uvicorn==0.27.1
